test-requirements.txt
coveralls
nose
nose-cov
coverage
pep8
mock
pyHamcrest
pyflakes
matchmock
